chore(gatsby): Migrate utils/page-data to TypeScript
This commit migrates the existing `src/utils/page-data.ts` file to
TypeScript by trying to stick as close as possible to the original
JavaScript file.

To take advantage of existing Redux typings, I tried to extend the
typings from the `IGatsbyPage` interface describing page data.
